requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
redis==5.0.8
celery==5.4.0
pdfplumber==0.11.4
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
statsmodels==0.14.2
<<<<<<< HEAD
groq==0.25.0
=======
anthropic==0.34.0
>>>>>>> 8a260015108003b30586cdffc9a5d14ec1407d73
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
<<<<<<< HEAD
google-generativeai>=0.5.0
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.0.0
=======
>>>>>>> 8a260015108003b30586cdffc9a5d14ec1407d73
